Increase version and fix invoice position save

+++ src/main/java/de/wichmann/sevdesk/model/ModelInvoice.java
@@ -37,7 +37,7 @@ public class ModelInvoice
 
 	@JsonProperty("id")
 
-	private Long id;
+	private Integer id;
 
 	@JsonProperty("invoiceNumber")
 
@@ -572,12 +572,12 @@ public class ModelInvoice
 		return header;
 	}
 
-	public Long getId()
+	public Integer getId()
 	{
 		return id;
 	}
 
-	public void setId(Long id)
+	public void setId(Integer id)
 	{
 		this.id = id;
 	}
